task1/requirements.txt
apache-airflow==2.10.5
psycopg2-binary==2.9.10
faker==36.1.1
requests==2.32.3
apache-airflow-providers-postgres==5.14.0
